fix: prisma error

+++ src/components/modules/Workspace/WorkspaceDelete/_action.tsx
@@ -3,11 +3,21 @@ import { revalidatePath } from "next/cache";
 import { db } from "@/server/db";
 import { getServerAuthSession } from "@/server/auth";
 import { redirect } from "next/navigation";
+import config from "@/constants/url.constant";
 
-export const deleteWorkspace = async (slug: string | null) => {
+export const deleteWorkspace = async (slug: string | null, id: number) => {
+  console.log(slug);
+
   try {
     const session = await getServerAuthSession();
     if (!session) return { error: "You must be logged in" };
+
+    const workspace = await db.workspace.findUnique({ where: { id } });
+    if (!workspace) return { error: "Workspace not found" };
+
+    await db.workspaceMember.deleteMany({
+      where: { workspaceId: workspace.id },
+    });
 
     await db.workspace.deleteMany({
       where: { slug },
@@ -17,6 +27,5 @@ export const deleteWorkspace = async (slug: string | null) => {
     return { error: error.message };
   }
 
-  revalidatePath("/dashboard");
-  redirect("/dashboard");
+  revalidatePath(config.url.WORKSPACE_URL());
 };
